add edit user, profil, home for superadmin

+++ w3/LibrAspire2/routes/web.php
@@ -13,18 +13,16 @@ use App\Http\Controllers\PeminjamanController;
 |--------------------------------------------------------------------------
 */
 
-// Landing page
 Route::get('/', function () {
     return view('landingpage');
-});
+})->name('landingpage');
 
-// Login & Register
 Route::get('/login', [AuthController::class, 'loginForm'])->name('login');
 Route::post('/login', [AuthController::class, 'login']);
 
 Route::get('/register', function () {
     return view('register');
-});
+})->name('register');
 Route::post('/register', [AuthController::class, 'register']);
 
 
@@ -35,28 +33,23 @@ Route::post('/register', [AuthController::class, 'register']);
 */
 Route::middleware('auth')->group(function () {
 
-    Route::get('/logout', [AuthController::class, 'logout']);
+    Route::get('/logout', [AuthController::class, 'logout'])->name('logout');
+
+    Route::get('/dashboard', function () {
+        $user = Auth::user();
+        if ($user->role == 'admin') return redirect('/admin');
+        if ($user->role == 'member') return redirect('/member');
+        if ($user->role == 'superadmin') return redirect('/superadmin');
+    })->name('dashboard');
+
 
     /*
     |--------------------------------------------------------------------------
-    | REDIRECT OTOMATIS SESUAI ROLE
+    | USER PROFILE (SEMUA ROLE)
     |--------------------------------------------------------------------------
     */
-    Route::get('/dashboard', function () {
-        $user = Auth::user();
-
-        if ($user->role == 'admin') {
-            return redirect('/admin');
-        }
-
-        if ($user->role == 'member') {
-            return redirect('/member');
-        }
-
-        if ($user->role == 'superadmin') {
-            return redirect('/superadmin');
-        }
-    });
+    Route::get('/profile/edit', [UserController::class, 'editProfile'])->name('profile.edit');
+    Route::post('/profile/update', [UserController::class, 'updateProfile'])->name('profile.update');
 
 
     /*
@@ -68,26 +61,26 @@ Route::middleware('auth')->group(function () {
 
         Route::get('/', function () {
             return view('admin.home');
-        });
+        })->name('admin.home');
 
-        Route::get('/profil', function () {
+        // Buku Admin
+        Route::get('/buku', [BukuController::class, 'index'])->name('admin.buku.index');
+        Route::get('/buku/create', [BukuController::class, 'create'])->name('admin.buku.create');
+        Route::post('/buku', [BukuController::class, 'store'])->name('admin.buku.store');
+        Route::get('/buku/{id}/edit', [BukuController::class, 'edit'])->name('admin.buku.edit');
+        Route::put('/buku/{id}', [BukuController::class, 'update'])->name('admin.buku.update');
+        Route::delete('/buku/{id}', [BukuController::class, 'destroy'])->name('admin.buku.destroy');
+
+        // Peminjaman Admin
+        Route::get('/peminjaman', [PeminjamanController::class, 'index'])->name('admin.peminjaman.index');
+
+        Route::get('/profile', function () {
             return view('admin.profile');
-        });
+        })->name('admin.profile');
 
-        Route::get('/profil/edit', function () {
-            return view('admin.editprofile');
-        });
-
-        // buku
-        Route::get('/buku', [BukuController::class, 'index']);
-        Route::get('/buku/create', [BukuController::class, 'create']);
-        Route::post('/buku', [BukuController::class, 'store']);
-        Route::get('/buku/{id}/edit', [BukuController::class, 'edit']);
-        Route::put('/buku/{id}', [BukuController::class, 'update']);
-        Route::delete('/buku/{id}', [BukuController::class, 'destroy']);
-
-        // peminjaman
-        Route::get('/peminjaman', [PeminjamanController::class, 'index']);
+        Route::get('/profile/edit', function () {
+            return view('admin.editprofile', ['user' => Auth::user()]);
+        })->name('admin.editprofile');
     });
 
 
@@ -100,26 +93,27 @@ Route::middleware('auth')->group(function () {
 
         Route::get('/', function () {
             return view('member.home');
-        });
-
-        Route::get('/profil', function () {
-            return view('member.profile');
-        });
-
-        Route::get('/profil/edit', function () {
-            return view('member.editprofile');
-        });
+        })->name('member.home');
 
         Route::get('/contact', function () {
             return view('member.contactus');
-        });
+        })->name('member.contact');
 
-        Route::get('/buku', [BukuController::class, 'index']);
+        // Buku Member
+        Route::get('/buku', [BukuController::class, 'index'])->name('member.buku.index');
         Route::get('/buku/{id}', function () {
             return view('member.buku.detail');
-        });
+        })->name('member.buku.detail');
 
-        Route::post('/pinjam/{id}', [PeminjamanController::class, 'store']);
+        Route::post('/pinjam/{id}', [PeminjamanController::class, 'store'])->name('member.pinjam');
+        
+        Route::get('/profile', function () {
+            return view('member.profile');
+        })->name('member.profile');
+
+        Route::get('/profile/edit', function () {
+            return view('member.editprofile', ['user' => Auth::user()]);
+        })->name('member.editprofile');
     });
 
 
@@ -132,21 +126,25 @@ Route::middleware('auth')->group(function () {
 
         Route::get('/', function () {
             return view('superadmin.home');
-        });
+        })->name('superadmin.home');
 
-        Route::get('/profil', function () {
+        // MANAGE USER
+        Route::get('/user', [UserController::class, 'index'])->name('superadmin.user.index');
+
+        // update role 
+        Route::put('/user/{id}/role', [UserController::class, 'updateRole'])->name('superadmin.user.updateRole');
+
+        // delete user
+        Route::delete('/user/{id}', [UserController::class, 'destroy'])->name('superadmin.user.destroy');
+
+        Route::get('/profile', function () {
             return view('superadmin.profile');
-        });
+        })->name('superadmin.profile');
 
-        Route::get('/profil/edit', function () {
-            return view('superadmin.editprofile');
-        });
+        Route::get('/profile/edit', function () {
+            return view('superadmin.editprofile', ['user' => Auth::user()]);
+        })->name('superadmin.editprofile');
 
-        // manage user
-        Route::get('/user', [UserController::class, 'index']);
-        Route::get('/user/{id}/edit', [UserController::class, 'edit']);
-        Route::put('/user/{id}', [UserController::class, 'update']);
-        Route::delete('/user/{id}', [UserController::class, 'destroy']);
     });
 
 });
